moved sse stream to tripimwatching
