+++ client/components/TripImWatching.jsx
@@ -1,11 +1,36 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import ProgressBar from './ProgressBar';
 import TripViewingCard from './TripViewingCard';
 
 
 
-function TripImWatching(props) {
-    
+function TripImWatching({userInfo}) {
+        
+  //SSE - render trips
+  const [trips, setTrips] = useState([]);
+  useEffect(() => {
+    const source = new EventSource(`http://localhost:3000/stream/1234567890`, { //replace 123456789 with current user's phone_number
+      withCredentials: false,
+    }); //maybe need to add to webpack?
+
+    source.addEventListener('open', () => {
+      console.log('SSE opened!');
+    });
+
+    source.addEventListener('message', (e) => {
+      // console.log(e.data);
+      const data = JSON.parse(e.data);
+      setTrips(data);
+    });
+
+    source.addEventListener('error', (e) => {
+      console.error('Error: ', e);
+    });
+
+    return () => {
+      source.close();
+    };
+  }, []);
 
   return (
     <>
