+++ client/pages/Dashboard.jsx
@@ -24,32 +24,6 @@ function Dashboard({ userInfo }) {
     setActiveComponent(componentName);
   };
 
-  //SSE - render trips
-  const [trips, setTrips] = useState([]);
-  useEffect(() => {
-    const source = new EventSource(`http://localhost:3000/stream/1234567890`, { //replace 123456789 with current user's phone_number
-      withCredentials: false,
-    }); //maybe need to add to webpack?
-
-    source.addEventListener('open', () => {
-      console.log('SSE opened!');
-    });
-
-    source.addEventListener('message', (e) => {
-      // console.log(e.data);
-      const data = JSON.parse(e.data);
-      setTrips(data);
-    });
-
-    source.addEventListener('error', (e) => {
-      console.error('Error: ', e);
-    });
-
-    return () => {
-      source.close();
-    };
-  }, []);
-
   return (
     <div className="dashboard-container">
       {/* SSE - Render trips */}
@@ -69,7 +43,9 @@ function Dashboard({ userInfo }) {
           contacts={contacts} 
           setContacts={setContacts} 
         /> }
-        {activeComponent === 'tripsImWatching' && <TripImWatching />}
+        {activeComponent === 'tripsImWatching' && <TripImWatching
+          userInfo={userInfo} 
+        />}
         {activeComponent === 'chatPage' &&  <ChatPage 
           path= '/chat' 
           socket={socket}
